LayoutTests/platform/win/fast/events/popup-blocked-from-fake-user-gesture-expected.txt

Click Here
PASS win is undefined.
